apply changes to rest of sources

+++ src/providers/postgres/qgspostgresconn.h
@@ -130,7 +130,7 @@ struct QgsPostgresLayerProperty
     {
       if ( !typeString.isEmpty() )
         typeString += '|';
-      typeString += QString::number( type );
+      typeString += QString::number( static_cast<int>( type ) );
     }
     QString sridString;
     const auto constSrids = srids;
